task_delay_system/requirements.txt
Django==4.2.28
dj-database-url==3.1.0
gunicorn==25.0.3
mysqlclient==2.2.8
whitenoise==6.11.0
asgiref==3.11.1
sqlparse==0.5.5
packaging==26.0
python-decouple==3.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0
celery==5.3.6
django-celery-beat==2.5.0
redis==5.0.1
django-ratelimit==4.1.0
